chore: rename *.js to *.ts and update import reference
