+++ packages/babel-helpers/package.json
@@ -19,6 +19,7 @@
     "@babel/types": "workspace:^"
   },
   "devDependencies": {
+    "@babel/core": "workspace:^",
     "@babel/generator": "workspace:^",
     "@babel/helper-plugin-test-runner": "workspace:^",
     "@babel/helper-transform-fixture-test-runner": "workspace:^",
